refactor(service,adapter): split under 800-line ceiling (#137) (#144)
service.py 1019 -> 790 lines; adapter.py 1107 -> 761 lines.

service.py extracts:
- service_cv.py — compute_preview_splits, validate_inner_valid,
  default_strategy_for_task, default_cv_state, GROUP/TIME_STRATEGIES
- service_columns.py — detect_task, auto_configure_column,
  calc_feature_summary, merge_best_params_into_config

adapter.py extracts:
- adapter_internals.py — version guard, _serialize_*, deep_merge,
  get/set/unset_nested, extract_defaults, convert_metric_entries,
  enforce_auto_num_leaves
- adapter_results.py — render_plot, list_available_plots,
  render_inference_plot, task_for_model

Composition only (no inheritance). Backward-compatible
LizyMLAdapter._<helper> static aliases keep existing tests untouched.
All 928 pytest tests pass; ruff + ruff format + mypy strict clean.
